week5_ocr_basics_project(AliHashim)/requirements.txt
pytesseract
easyocr
opencv-python
pillow
pandas
matplotlib
numpy
